requirements.txt
Flask==2.2.5
flask-cors
pymongo
python-dotenv
pytesseract
easyocr
pdf2image
python-docx
transformers
torch
sentencepiece
gunicorn
